generate_ledger/pyproject.toml
[project]
name = "generate-ledger"
version = "0.0.4a1"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Michael Legleux", email = "legleux@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "base58>=2.1.1",
    "httpx>=0.28.1",
    "ruamel-yaml>=0.18.14",
    "xrpl-py>=4.2.0",
]

[project.scripts]
gen = "generate_ledger:main"
compose = "generate_ledger:compose"
ledger = "generate_ledger:ledger"
config = "generate_ledger:config"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
